I feel like highlighting the tag brackets may've made more sense with my old color scheme but, with Klere, I don't think it's needed.
The new ones were when I was trying to use AutoComplete mode and I
didn't like the color scheme.

I still don't but I don't really feel fully good with what I've come
up with, either, so I'll hold off on inflicting it on the users of the
theme, yet.